Separate bin operations from IFileIOHandler
