+++ src/FileSurfer/Models/Shell/IFileRestorer.cs
@@ -3,6 +3,18 @@
 public interface IFileRestorer
 {
     /// <summary>
+    /// Moves a file to the trash.
+    /// </summary>
+    /// <returns>A <see cref="IResult"/> representing the result of the operation and potential errors.</returns>
+    public IResult MoveFileToTrash(string filePath);
+
+    /// <summary>
+    /// Moves a directory to the trash.
+    /// </summary>
+    /// <returns>A <see cref="IResult"/> representing the result of the operation and potential errors.</returns>
+    public IResult MoveDirToTrash(string dirPath);
+
+    /// <summary>
     /// Restores a file based on <paramref name="originalFilePath"/>.
     /// </summary>
     /// <returns>A <see cref="IResult"/> representing the result of the operation and potential errors.</returns>
