+++ src/FileSurfer/Models/Shell/WindowsFileRestorer.cs
@@ -1,6 +1,8 @@
 using System;
 using System.IO;
 using System.Runtime.InteropServices;
+using FileSurfer.Models.FileInformation;
+using Microsoft.VisualBasic.FileIO;
 using Shell32;
 
 namespace FileSurfer.Models.Shell;
@@ -15,6 +17,15 @@ public class WindowsFileRestorer : IFileRestorer
     private const int NameColumn = 0;
     private const int PathColumn = 1;
     private const string RestoreVerb = "ESTORE";
+
+    private readonly long _showDialogLimit;
+    private readonly IFileInfoProvider _fileInfoProvider;
+
+    public WindowsFileRestorer(long showDialogLimit, IFileInfoProvider fileInfoProvider)
+    {
+        _fileInfoProvider = fileInfoProvider;
+        _showDialogLimit = showDialogLimit;
+    }
 
     public IResult RestoreFile(string originalFilePath) => RestoreEntry(originalFilePath);
 
@@ -59,6 +70,44 @@ public class WindowsFileRestorer : IFileRestorer
                 verbObject.DoIt();
                 return;
             }
+        }
+    }
+
+    public IResult MoveFileToTrash(string filePath)
+    {
+        try
+        {
+            FileSystem.DeleteFile(
+                filePath,
+                _fileInfoProvider.GetFileSizeB(filePath) > _showDialogLimit
+                    ? UIOption.AllDialogs
+                    : UIOption.OnlyErrorDialogs,
+                RecycleOption.SendToRecycleBin,
+                UICancelOption.ThrowException
+            );
+            return SimpleResult.Ok();
+        }
+        catch (Exception ex)
+        {
+            return SimpleResult.Error(ex.Message);
+        }
+    }
+
+    public IResult MoveDirToTrash(string dirPath)
+    {
+        try
+        {
+            FileSystem.DeleteDirectory(
+                dirPath,
+                UIOption.AllDialogs,
+                RecycleOption.SendToRecycleBin,
+                UICancelOption.ThrowException
+            );
+            return SimpleResult.Ok();
+        }
+        catch (Exception ex)
+        {
+            return SimpleResult.Error(ex.Message);
         }
     }
 }
