+++ src/FileSurfer/ViewModels/MainWindowViewModel.cs
@@ -42,6 +42,7 @@ public sealed class MainWindowViewModel : ReactiveObject, IDisposable
     private string ThisPCLabel => FileSurferSettings.ThisPCLabel;
 
     private readonly IFileIOHandler _fileIOHandler;
+    private readonly IFileRestorer _fileRestorer;
     private readonly IFileInfoProvider _fileInfoProvider;
     private readonly IFileProperties _fileProperties;
     private readonly IShellHandler _shellHandler;
@@ -214,6 +215,7 @@ public sealed class MainWindowViewModel : ReactiveObject, IDisposable
     public MainWindowViewModel(
         string initialDir,
         IFileIOHandler fileIOHandler,
+        IFileRestorer fileRestorer,
         IFileProperties fileProperties,
         IFileInfoProvider fileInfoProvider,
         IShellHandler shellHandler,
@@ -222,6 +224,7 @@ public sealed class MainWindowViewModel : ReactiveObject, IDisposable
     )
     {
         _fileIOHandler = fileIOHandler;
+        _fileRestorer = fileRestorer;
         _fileProperties = fileProperties;
         _fileInfoProvider = fileInfoProvider;
         _shellHandler = shellHandler;
@@ -1189,6 +1192,7 @@ public sealed class MainWindowViewModel : ReactiveObject, IDisposable
     public void MoveToTrash()
     {
         MoveFilesToTrash operation = new(
+            _fileRestorer,
             _fileIOHandler,
             SelectedFiles.ConvertToArray(entry => entry.FileSystemEntry)
         );
